% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.tools-package.R
\docType{package}
\name{taxonomy.tools-package}
\alias{taxonomy.tools}
\alias{taxonomy.tools-package}
\title{taxonomy.tools: Tools to Work with Taxonomic Data}
\description{
Tools to Work with Taxonomic Data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Pakillo/taxonomy.tools}
  \item \url{https://pakillo.github.io/taxonomy.tools/}
  \item Report bugs at \url{https://github.com/Pakillo/taxonomy.tools/issues}
}

}
\author{
\strong{Maintainer}: Francisco Rodriguez-Sanchez \email{f.rodriguez.sanc@gmail.com} (\href{https://orcid.org/0000-0002-7981-1599}{ORCID})

}
\keyword{internal}
