% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcvp.R
\name{wcvp_match_names_parallel}
\alias{wcvp_match_names_parallel}
\title{Parallelised matching of plant taxonomic names against WCVP}
\usage{
wcvp_match_names_parallel(
  df = NULL,
  name_col = NULL,
  author_col = NULL,
  fuzzy = TRUE,
  plan = "multisession",
  cores = 1,
  ...
)
}
\arguments{
\item{df}{Data frame containing taxonomic names to be matched against WCVP.}

\item{name_col}{Character. The column in 'df' containing the taxon names for matching.}

\item{author_col}{Character. The column in 'df' that has the name authority, to aid matching.
Set to NULL to match with no author string.}

\item{fuzzy}{Logical; whether or not fuzzy matching should be used for names that could not be matched exactly.}

\item{plan}{Character. Name of a \code{\link[future:plan]{future::plan()}} to be used for parallelisation.
Recommended plan is "multisession". Use "sequential" for non-parallel processing.}

\item{cores}{Number of cores to use in parallel.}

\item{...}{Further arguments to be passed to \code{\link[rWCVP:wcvp_match_names]{rWCVP::wcvp_match_names()}}}
}
\value{
A data frame
}
\description{
Parallelised version of \code{\link[rWCVP:wcvp_match_names]{rWCVP::wcvp_match_names()}} to check plant taxonomic names
against the World Checklist of Vascular Plants (WCVP), using \code{\link[future.apply:future.apply]{future.apply::future.apply()}}.
}
\examples{
\dontrun{
df <- data.frame(taxon = c("Laurus nobilis", "Laurus nobilis", "Laurus nobili"),
                  author = c(NA, "L.", NA))
out <- wcvp_match_names_parallel(df, name_col = "taxon", author_col = "author",
  cores = 3)
}
}
