% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcvp.R
\name{wcvp_resolve_multiple_matches}
\alias{wcvp_resolve_multiple_matches}
\title{Resolve multiple matches in WCVP}
\usage{
wcvp_resolve_multiple_matches(df = NULL, name_col = NULL)
}
\arguments{
\item{df}{Data frame with matching results, as produced by \code{\link[rWCVP:wcvp_match_names]{rWCVP::wcvp_match_names()}} or
\code{\link[=wcvp_match_names_parallel]{wcvp_match_names_parallel()}}}

\item{name_col}{Character. Name of the column in 'df' containing the taxon names.}
}
\value{
A data frame
}
\description{
This function tries to resolve automatically the cases of multiple matches found
after running a vector of taxonomic names against the World Checklist of Vascular Plants (WCVP).
}
\examples{
\dontrun{
df <- data.frame(taxon = "Acacia dealbata")
matching <- wcvp_match_names_parallel(df, name_col = "taxon")
out <- wcvp_resolve_multiple_matches(matching, name_col = "taxon")
}
}
\seealso{
https://matildabrown.github.io/rWCVP/articles/redlist-name-matching.html
}
\author{
Adapted from \code{rWCVP} \href{https://matildabrown.github.io/rWCVP/articles/redlist-name-matching.html#multiple-matches}{vignette}
}
